Add the possibility to set the release name directly as makefile parameter
With this patch, make tarball and make export still work as expected,
but if you specify RELEASENAME=foo, then the directories and tarballs
are named flashrom-foo instead of flashrom-0.9.0-r703. This makes
release creation a lot easier. As an example, look at creating the 0.9.1
tarball: # make tarball RELEASENAME=0.9.1

Corresponding to flashrom svn r704.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Stefan Reinauer <stepan@coresystems.de>