Refine selection of the default programmer
If neither internal (old default) nor dummy (safe default) programmer
are selected, we must pick a sensible default programmer.

Since there is no reason to prefer a particular external programmer,
we abort compilation if more than one of them is selected. If only one
is selected, it is clear that the user wants that one to become the
default. This fixes single-programmer compilation.

Corresponding to flashrom svn r858.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Sean Nelson <audiohacked@gmail.com>