Makefile: detect unsupported configuration options
Detect incompatible CONFIG_FOO=yes for the specified target and refuse
to compile with a meaningful error message. Set the default for
incompatible CONFIG_FOO to no.

Just running "make" should result in a a build which compiles the common
subset of available and working features.

Corresponding to flashrom svn r1048.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Idwer Vollering <vidwer+lists.flashrom@gmail.com>